Create new posts app and move post models from utils
